pkg/asset/installconfig/aws/platform: Put Platform in its own file
"aws.go" doesn't add any information that is not covered in the
package name.  Putting Platform in its own platform.go file will make
it easier to find for folks tab-completing into the directory in
search of Platform.
